// example/generated/flow/post/selects.ts
// @generated by next-prisma-flow-state-engine
// file: selects.ts

import { Prisma } from "@prisma/client";

export const PostDeepSelect = {
  id: true,
  title: true,
  content: true,
  published: true,
  views: true,
  createdAt: true,
  updatedAt: true,
  authorId: true,
  author: {
    select: {
      id: true,
      email: true,
      name: true,
      bio: true,
      createdAt: true,
      updatedAt: true,
      posts: false,
      comments: {
        take: 50,
        select: {
          id: true,
          content: true,
          createdAt: true,
          updatedAt: true,
          postId: true,
          authorId: true,
          author: false,
          parentId: true,
          replies: {
            take: 50,
            select: {
              id: true,
              content: true,
              createdAt: true,
              updatedAt: true,
              postId: true,
              authorId: true,
              author: false,
              parentId: true,
              replies: false,
            },
          },
        },
      },
    },
  },
  comments: {
    take: 50,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      author: {
        select: {
          id: true,
          email: true,
          name: true,
          bio: true,
          createdAt: true,
          updatedAt: true,
          posts: false,
          comments: false,
        },
      },
      parentId: true,
      replies: {
        take: 50,
        select: {
          id: true,
          content: true,
          createdAt: true,
          updatedAt: true,
          postId: true,
          authorId: true,
          author: {
            select: {
              id: true,
              email: true,
              name: true,
              bio: true,
              createdAt: true,
              updatedAt: true,
              posts: false,
              comments: false,
            },
          },
          parentId: true,
          replies: false,
        },
      },
    },
  },
  tags: {
    take: 50,
    select: {
      id: true,
      name: true,
      createdAt: true,
    },
  },
} as const satisfies Prisma.PostSelect;

export const PostShallowSelect = {
  id: true,
  title: true,
  content: true,
  published: true,
  views: true,
  createdAt: true,
  updatedAt: true,
  authorId: true,
} as const satisfies Prisma.PostSelect;

export const PostListSelect = {
  id: true,
  title: true,
  content: true,
  published: true,
  views: true,
  createdAt: true,
  updatedAt: true,
  authorId: true,
  author: {
    select: {
      id: true,
      email: true,
      name: true,
      bio: true,
      createdAt: true,
      updatedAt: true,
      organizationId: true,
    },
  },
  comments: {
    take: 10,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      author: {
        select: {
          id: true,
          email: true,
          name: true,
          bio: true,
          createdAt: true,
          updatedAt: true,
          organizationId: true,
        },
      },
      parentId: true,
      organizationId: true,
    },
  },
  tags: {
    take: 10,
    select: {
      id: true,
      name: true,
      createdAt: true,
      organizationId: true,
    },
  },
} as const satisfies Prisma.PostSelect;