// example/generated/flow/author/selects.ts
// @generated by next-prisma-flow-state-engine
// file: selects.ts

import { Prisma } from "@prisma/client";

export const AuthorDeepSelect = {
  id: true,
  email: true,
  name: true,
  bio: true,
  createdAt: true,
  updatedAt: true,
  posts: {
    take: 50,
    select: {
      id: true,
      title: true,
      content: true,
      published: true,
      views: true,
      createdAt: true,
      updatedAt: true,
      authorId: true,
      author: false,
      comments: {
        take: 50,
        select: {
          id: true,
          content: true,
          createdAt: true,
          updatedAt: true,
          postId: true,
          authorId: true,
          author: false,
          parentId: true,
          replies: {
            take: 50,
            select: {
              id: true,
              content: true,
              createdAt: true,
              updatedAt: true,
              postId: true,
              authorId: true,
              author: false,
              parentId: true,
              replies: false,
            },
          },
        },
      },
      tags: {
        take: 50,
        select: {
          id: true,
          name: true,
          createdAt: true,
        },
      },
    },
  },
  comments: {
    take: 50,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      author: false,
      parentId: true,
      replies: {
        take: 50,
        select: {
          id: true,
          content: true,
          createdAt: true,
          updatedAt: true,
          postId: true,
          authorId: true,
          author: false,
          parentId: true,
          replies: false,
        },
      },
    },
  },
} as const satisfies Prisma.AuthorSelect;

export const AuthorShallowSelect = {
  id: true,
  email: true,
  name: true,
  bio: true,
  createdAt: true,
  updatedAt: true,
} as const satisfies Prisma.AuthorSelect;

export const AuthorListSelect = {
  id: true,
  email: true,
  name: true,
  bio: true,
  createdAt: true,
  updatedAt: true,
  posts: {
    take: 10,
    select: {
      id: true,
      title: true,
      content: true,
      published: true,
      views: true,
      createdAt: true,
      updatedAt: true,
      authorId: true,
      organizationId: true,
    },
  },
  comments: {
    take: 10,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      author: {
        select: {
          id: true,
          email: true,
          name: true,
          bio: true,
          createdAt: true,
          updatedAt: true,
          organizationId: true,
        },
      },
      parentId: true,
      organizationId: true,
    },
  },
} as const satisfies Prisma.AuthorSelect;