// example/generated/flow/comment/selects.ts
// @generated by next-prisma-flow-state-engine
// file: selects.ts

import { Prisma } from "@prisma/client";

export const CommentDeepSelect = {
  id: true,
  content: true,
  createdAt: true,
  updatedAt: true,
  postId: true,
  authorId: true,
  author: {
    select: {
      id: true,
      email: true,
      name: true,
      bio: true,
      createdAt: true,
      updatedAt: true,
      posts: {
        take: 50,
        select: {
          id: true,
          title: true,
          content: true,
          published: true,
          views: true,
          createdAt: true,
          updatedAt: true,
          authorId: true,
          author: false,
          comments: false,
          tags: {
            take: 50,
            select: {
              id: true,
              name: true,
              createdAt: true,
            },
          },
        },
      },
      comments: false,
    },
  },
  parentId: true,
  replies: {
    take: 50,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      author: {
        select: {
          id: true,
          email: true,
          name: true,
          bio: true,
          createdAt: true,
          updatedAt: true,
          posts: {
            take: 50,
            select: {
              id: true,
              title: true,
              content: true,
              published: true,
              views: true,
              createdAt: true,
              updatedAt: true,
              authorId: true,
              author: false,
              comments: false,
              tags: {
                take: 50,
                select: {
                  id: true,
                  name: true,
                  createdAt: true,
                },
              },
            },
          },
          comments: false,
        },
      },
      parentId: true,
      replies: false,
    },
  },
} as const satisfies Prisma.CommentSelect;

export const CommentShallowSelect = {
  id: true,
  content: true,
  createdAt: true,
  updatedAt: true,
  postId: true,
  authorId: true,
  parentId: true,
} as const satisfies Prisma.CommentSelect;

export const CommentListSelect = {
  id: true,
  content: true,
  createdAt: true,
  updatedAt: true,
  postId: true,
  authorId: true,
  author: {
    select: {
      id: true,
      email: true,
      name: true,
      bio: true,
      createdAt: true,
      updatedAt: true,
      organizationId: true,
    },
  },
  parentId: true,
  replies: {
    take: 10,
    select: {
      id: true,
      content: true,
      createdAt: true,
      updatedAt: true,
      postId: true,
      authorId: true,
      parentId: true,
      organizationId: true,
    },
  },
} as const satisfies Prisma.CommentSelect;